requirements.txt
discord.py>=1.1
git+https://github.com/mathsman5133/coc.py@events
asyncpg